Use veewee and sahara gems from github.
The current published versions of veewee and sahara rely on the Vagrant 1.0.x gem, which is not present for versions 1.1.x and later.  The Vagrant 1.0.x gem may conflict and cause issues with more modern versions of Vagrant.

Since the version of veewee on github does not have this requirement, we should use that instead for now.  When a new version is published, we should switch to that version.

The version of the sahara gem on github still relies on Vagrant 1.0.x to function properly.  There is a fork that supports Vagrant 1.1.x.  We should use that until the master version of the gem supports Vagrant 1.1.x.